New About feature
I have added a new function so that you have the “Info” item first in the menu, from which two more items branch off, once to the list of all contributors and once to the version field. The version number is stored in the file. If this does not match the latest version number of the last release, it will indicate that a new update is available.