Adds Generative_Code
This is the program that the main text editor can pull on to get the outputs for the generative model. This module is currently not finished and under testing. *If anyone can help edit this it will be great because I am not with my desktop computer so I can't run the code.*